+++ src/main/java/gov/nasa/pds/harvest/meta/BasicMetadataExtractor.java
@@ -17,6 +17,7 @@ import org.w3c.dom.NodeList;
 import gov.nasa.pds.harvest.Constants;
 import gov.nasa.pds.harvest.cfg.model.Configuration;
 import gov.nasa.pds.harvest.util.PdsStringUtils;
+import gov.nasa.pds.registry.common.meta.Metadata;
 import gov.nasa.pds.registry.common.util.xml.XPathUtils;
